+++ src/renderer/core/layout/useTransformState.ts
@@ -51,7 +51,7 @@
  */
 import { computed, reactive, readonly } from 'vue'
 
-import type { LGraphCanvas } from '../../lib/litegraph/src/litegraph'
+import type { LGraphCanvas } from '@/lib/litegraph/src/litegraph'
 
 export interface Point {
   x: number
