+++ src/renderer/extensions/minimap/MiniMap.vue
