+++ src/renderer/extensions/minimap/MiniMapPanel.vue
